src/pkgw_pitchy_power_law.c: incorporate the pitch-angle dependence
Tadaa! Now that we have the infrastructure, I think this is all that's needed
to have a distribution function with a pitch-angle dependence. With no
pitch-angle dependence, the equivalent term would just be unity integrated
from -1 to 1, so the integral of the pitch angle term should be 2. The
expression `sin^2 (cosxi^2 * pi)` has the offset structure I want; its
integral can be evaluated numerically to something gross that I've put in as a
constant. Note that `sin^2 (cosxi * pi)` has a *much* simpler integral, but
the square term shifts the peaks of emphasis to more extreme pitch angles in a
way that I think is good.