src/integrator/integrate.c: use different integration strategy for pitchy Stokes V
The Stokes V integrals that Symphony does end up looking kind of like a cycle
of a sine wave, with a total close to zero each time. Apparently this causes
problems for the integrator, so Symphony integrates each lobe separately and
adds up the two results.

The anisotropic "pitchy" power law breaks the symmetry of this sine-like
curve, and I have a theory that this is leading to bad integration results for
pitchy Stokes V. As a first attempt, let's not use the two-lobe approach for
the pitchy case.