odyssey: our parameter space is really just 3D
J and alpha just scale linearly with n_e, so we don't really need to explore
it as a parameter. The physics of the synchrotron emission then depend almost
exclusively on nu/nu_cyc or nu/B, except that J and alpha scale up and down
with nu, respectively. So if we fix a reference nu and n_e, we can derive
coefficients using just a 3D parameter space: s, theta, and p (in the
power-law distribution case). That's a drop from 5D to 3D!